--- livekit/src/room/data_stream/incoming/reader.rs
@@ -1,182 +1,0 @@
-// Copyright 2025 LiveKit, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use super::manager::IncomingChunk;
-use crate::data_stream::info::AnyStreamInfo;
-use crate::data_stream::{
-    ByteStreamInfo, StreamError, StreamProgress, StreamResult, TextStreamInfo,
-};
-use bytes::{Bytes, BytesMut};
-use futures_util::{Stream, StreamExt};
-use std::pin::Pin;
-use std::task::{Context, Poll};
-use tokio::sync::mpsc::{self, UnboundedReceiver, UnboundedSender};
-
-/// Reader for an incoming data stream.
-pub trait StreamReader: Stream<Item = StreamResult<(Self::Output, StreamProgress)>> {
-    /// Type of output this reader produces.
-    type Output;
-
-    /// Information about the underlying data stream.
-    type Info;
-
-    /// Returns a reference to the stream info.
-    fn info(&self) -> &Self::Info;
-
-    /// Reads all incoming chunks from the byte stream, concatenating them
-    /// into a single value which is returned once the stream closes normally.
-    ///
-    /// Returns the data consisting of all concatenated chunks.
-    ///
-    fn read_all(self) -> impl std::future::Future<Output = StreamResult<Self::Output>> + Send;
-}
-
-pub(super) type ChunkSender = UnboundedSender<StreamResult<IncomingChunk>>;
-type ChunkReceiver = UnboundedReceiver<StreamResult<IncomingChunk>>;
-
-/// Reader for an incoming byte data stream.
-pub struct ByteStreamReader {
-    info: ByteStreamInfo,
-    rx: ChunkReceiver,
-}
-
-impl StreamReader for ByteStreamReader {
-    type Output = Bytes;
-    type Info = ByteStreamInfo;
-
-    fn info(&self) -> &ByteStreamInfo {
-        &self.info
-    }
-
-    async fn read_all(mut self) -> StreamResult<Bytes> {
-        let mut buffer = BytesMut::new();
-        while let Some(result) = self.next().await {
-            match result {
-                Ok((bytes, _)) => buffer.extend_from_slice(&bytes),
-                Err(e) => return Err(e),
-            }
-        }
-        Ok(buffer.freeze())
-    }
-}
-
-impl ByteStreamReader {
-    /// Reads incoming chunks from the byte stream, writing them to a file as they are received.
-    ///
-    /// Parameters:
-    ///   - directory: The directory to write the file in. The system temporary directory is used if not specified.
-    ///   - name_override: The name to use for the written file, overriding stream name.
-    ///
-    /// Returns: The path of the written file on disk.
-    ///
-    pub async fn write_to_file(
-        mut self,
-        directory: Option<impl AsRef<std::path::Path>>,
-        name_override: Option<&str>,
-    ) -> StreamResult<std::path::PathBuf> {
-        let directory =
-            directory.map(|d| d.as_ref().to_path_buf()).unwrap_or_else(|| std::env::temp_dir());
-        let name = name_override.unwrap_or_else(|| &self.info.name);
-        let file_path = directory.join(name);
-
-        let mut file = tokio::fs::File::create(&file_path).await.map_err(StreamError::Io)?;
-
-        while let Some(result) = self.next().await {
-            let (bytes, _) = result?;
-            tokio::io::AsyncWriteExt::write_all(&mut file, &bytes)
-                .await
-                .map_err(StreamError::Io)?;
-        }
-        tokio::io::AsyncWriteExt::flush(&mut file).await.map_err(StreamError::Io)?;
-
-        Ok(file_path)
-    }
-}
-
-impl Stream for ByteStreamReader {
-    type Item = StreamResult<(Bytes, StreamProgress)>;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let this = self.get_mut();
-        match Pin::new(&mut this.rx).poll_recv(cx) {
-            Poll::Ready(Some(Ok(chunk))) => Poll::Ready(Some(Ok((chunk.content, chunk.progress)))),
-            Poll::Ready(Some(Err(e))) => Poll::Ready(Some(Err(e))),
-            Poll::Ready(None) => Poll::Ready(None),
-            Poll::Pending => Poll::Pending,
-        }
-    }
-}
-
-/// Reader for an incoming text data stream.
-pub struct TextStreamReader {
-    info: TextStreamInfo,
-    rx: ChunkReceiver,
-}
-
-impl StreamReader for TextStreamReader {
-    type Output = String;
-    type Info = TextStreamInfo;
-
-    fn info(&self) -> &TextStreamInfo {
-        &self.info
-    }
-
-    async fn read_all(mut self) -> StreamResult<String> {
-        let mut result = String::new();
-        while let Some(chunk) = self.next().await {
-            match chunk {
-                Ok((text, _)) => result.push_str(&text),
-                Err(e) => return Err(e),
-            }
-        }
-        Ok(result)
-    }
-}
-
-impl Stream for TextStreamReader {
-    type Item = StreamResult<(String, StreamProgress)>;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let this = self.get_mut();
-        match Pin::new(&mut this.rx).poll_recv(cx) {
-            Poll::Ready(Some(Ok(chunk))) => match String::from_utf8(chunk.content.into()) {
-                Ok(content) => Poll::Ready(Some(Ok((content, chunk.progress)))),
-                Err(e) => {
-                    this.rx.close();
-                    Poll::Ready(Some(Err(StreamError::from(e))))
-                }
-            },
-            Poll::Ready(Some(Err(e))) => Poll::Ready(Some(Err(e))),
-            Poll::Ready(None) => Poll::Ready(None),
-            Poll::Pending => Poll::Pending,
-        }
-    }
-}
-
-pub enum AnyStreamReader {
-    Byte(ByteStreamReader),
-    Text(TextStreamReader),
-}
-
-impl AnyStreamReader {
-    /// Creates a stream reader for the stream with the given info.
-    pub(super) fn from(info: AnyStreamInfo) -> (Self, ChunkSender) {
-        let (tx, rx) = mpsc::unbounded_channel();
-        let reader = match info {
-            AnyStreamInfo::Byte(info) => Self::Byte(ByteStreamReader { info, rx }),
-            AnyStreamInfo::Text(info) => Self::Text(TextStreamReader { info, rx }),
-        };
-        return (reader, tx);
-    }
-}
